Test generalized expression interpreter
