+++ src/test/scala/edu/colorado/plv/cuanto/scoot/interpreter/control/ControlInterpreterSpec.scala
@@ -1,0 +1,12 @@
+package edu.colorado.plv.cuanto.scoot.interpreter
+package control
+
+import org.scalatest.{FlatSpec, Matchers}
+import org.scalatest.prop.PropertyChecks
+
+////////////////////////////////////////////////////////////////////////
+
+class ControlInterpreterSpec extends FlatSpec with Matchers
+    with PropertyChecks {
+  // import Builder._
+}
