+++ src/test/scala/edu/colorado/plv/cuanto/scoot/interpreter/expression/ExpressionInterpreterSpec.scala
@@ -1,0 +1,51 @@
+package edu.colorado.plv.cuanto.scoot.interpreter
+package expression
+
+import org.scalatest.{FlatSpec, Matchers}
+import org.scalatest.prop.PropertyChecks
+
+////////////////////////////////////////////////////////////////////////
+
+class ExpressionInterpreterSpec extends FlatSpec with Matchers
+    with PropertyChecks {
+  import Builder._
+
+  val va = local("va")
+  val vb = local("vb")
+
+  val testEnv = Map(("va",IntR(3)),("vb",IntR(15)))
+
+  val exprTests = Table(
+    "expression" -> "denotation",
+    int(1) -> 1,
+    neg(int(1)) -> -1,
+    add(int(1))(int(1)) -> 2,
+    sub(int(3))(int(1)) -> 2,
+    mul(int(3))(int(2)) -> 6,
+    div(int(8))(int(4)) -> 2
+  )
+
+  val exprLocalTests = Table(
+    "expression" -> "denotation",
+    add(va)(int(1)) -> 4,
+    div(vb)(va) -> 5
+  )
+
+  "The Scoot interpreter" should "interpret stateless Values" in {
+    forAll (exprTests) { (e, n) =>
+      interpret(e,testEnv) should equal (Some(IntR(n)))
+    }
+  }
+
+  it should "interpret Values containing in-scope Locals" in {
+    forAll (exprLocalTests) { (e, n) =>
+      interpret(e,testEnv) should equal (Some(IntR(n)))
+    }
+  }
+
+  it should "give None when asked to interpret undefined Locals" in {
+    forAll (exprLocalTests) { (e, n) => interpret(e) should equal (None) }
+  }
+
+
+}
